Double precision camera, camera plugin, file dialog, non-coherent staging buffers
Co-authored: Nia Bickford <nbickford@nvidia.com>

Script sync from SHA: bb09c52e1a21976db3c58b368fb98babdf22ad70

+++ nvapp/elem_camera.cpp
@@ -22,7 +22,7 @@
 
 #include "elem_camera.hpp"
 
-float nvapp::ElementCamera::wheelSpeed = -10.0f;
+double nvapp::ElementCamera::wheelSpeed = -10.0;
 
 void nvapp::ElementCamera::updateCamera(std::shared_ptr<nvutils::CameraManipulator> m_cameraManip, ImGuiWindow* viewportWindow)
 {
